# swagger.yaml
swagger: "2.0"
info:
  description: "Sistema para gerenciar empréstimos em grupo em um banco de investimentos."
  version: "1.0.0"
  title: "Emprestimo Bancário"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "carlosebrand@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:44373"
basePath: "/api"
tags:
- name: "IATEC"
  description: "Teste desenvolvedor"
  externalDocs:
    description: "Find out more"
    url: "https://iatec.com/"
- name: "emprestimo"
  description: "Ações relativas à emprestimos bancarios"

schemes:
- "https"
- "http"
paths:
  /emprestimo:
    get:
      tags:
      - "emprestimo"
      summary: "Lista todos os emprestimos"
      description: "Lista todos os emprestimos"
      operationId: "getInventory"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            additionalProperties:
              type: "integer"
              format: "int32"
      security:
      - api_key: []

  /emprestimo/novo:
    post:
      tags:
      - "emprestimo"
      summary: "Cria um novo emprestimo"
      description: ""
      operationId: "placeOrder"
      produces:
        - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "quantia e linha de crédito"
        required: true
        schema:
          $ref: "#/definitions/Order"
      responses:
        "201":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        "400":
          description: "Invalid Order"
  
  /emprestimo/{id}:
    get:
      tags:
      - "emprestimo"
      summary: "Lista empréstimos por ID"
      description: "Para um response válido deve se passar um número inteiro"
      operationId: "getOrderById"
      produces:
        - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id de um emprestimo específico"
        required: true
        type: "integer"
        maximum: 10.0
        minimum: 1.0
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"
  
  /emprestimo/{id}/aumentar:
    put:
      tags:
      - "emprestimo"
      summary: "Aumenta o valor do emprestimo"
      description: ""
      operationId: "Aumentar"
      produces:
        - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id de um emprestimo específico"
        required: true
        type: "integer"
        maximum: 10.0
        minimum: 1.0
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"
    
  /emprestimo/{id}/pagar:
    put:
      tags:
      - "emprestimo"
      summary: "Reduz o valor do emprestimo"
      description: ""
      operationId: "Diminuir"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id de um emprestimo específico"
        required: true
        type: "integer"
        maximum: 10.0
        minimum: 1.0
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"    
  
  /linha-de-credito:
    post:
      tags:
      - "linha-de-credito"
      summary: "Cria uma linha de crédito"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
  
  /linha-de-credito/listar:
    get:
      tags:
      - "linha-de-credito"
      summary: "Lista todas as linhas de crédito"
      description: ""
      operationId: "loginUser"
      produces:
        - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "string"
          headers:
            X-Rate-Limit:
              type: "integer"
              format: "int32"
              description: "calls per hour allowed by the user"
            X-Expires-After:
              type: "string"
              format: "date-time"
              description: "date in UTC when token expires"
        "400":
          description: "Invalid username/password supplied"
  
  /linha-de-credito/{id}:
    get:
      tags:
      - "linha-de-credito"
      summary: "Lista uma linha de crédito especifica "
      description: ""
      operationId: "logoutUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id de um emprestimo específico"
        required: true
        type: "integer"
        maximum: 10.0
        minimum: 1.0
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"   
          
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Order:
    type: "object"
    properties:
      quantia:
        type: "integer"
        format: "int64"
      linhaDeCreditoId:
        type: "integer"
        format: "int64"
    xml:
      name: "Order"
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
    xml:
      name: "User"
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
  Pet:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "pet status in the store"
        enum:
        - "available"
        - "pending"
        - "sold"
    xml:
      name: "Pet"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"